// frontend/src/app/api/generated/model/cancellation-result-dto.ts
/**
 * Zana HMIS API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CreditNoteDto } from './credit-note-dto';


export interface CancellationResultDto { 
    billUid?: string;
    billStatus?: string;
    creditNote?: CreditNoteDto;
}