// frontend/src/app/api/generated/model/credit-note-dto.ts
/**
 * Zana HMIS API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { MoneyDto } from './money-dto';


export interface CreditNoteDto { 
    uid?: string;
    no?: string;
    patientUid?: string;
    amount?: MoneyDto;
    reference?: string;
    status?: string;
    patientBillUid?: string;
    businessDayUid?: string;
}