// frontend/src/app/api/generated/model/patient-bill-dto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { MoneyDto } from './money-dto';


export interface PatientBillDto { 
    uid?: string;
    patientUid?: string;
    billItem?: string;
    description?: string;
    kind?: PatientBillDto.KindEnum;
    qty?: number;
    amount?: MoneyDto;
    paid?: MoneyDto;
    balance?: MoneyDto;
    status?: PatientBillDto.StatusEnum;
    paymentType?: PatientBillDto.PaymentTypeEnum;
    membershipNo?: string;
    planUid?: string;
    businessDayUid?: string;
}
export namespace PatientBillDto {
    export const KindEnum = {
        Registration: 'REGISTRATION',
        Consultation: 'CONSULTATION',
        LabTest: 'LAB_TEST',
        Medicine: 'MEDICINE',
        Procedure: 'PROCEDURE',
        Radiology: 'RADIOLOGY',
        Ward: 'WARD'
    } as const;
    export type KindEnum = typeof KindEnum[keyof typeof KindEnum];
    export const StatusEnum = {
        Unpaid: 'UNPAID',
        Verified: 'VERIFIED',
        Covered: 'COVERED',
        Paid: 'PAID',
        None: 'NONE',
        Canceled: 'CANCELED'
    } as const;
    export type StatusEnum = typeof StatusEnum[keyof typeof StatusEnum];
    export const PaymentTypeEnum = {
        Cash: 'CASH',
        Insurance: 'INSURANCE'
    } as const;
    export type PaymentTypeEnum = typeof PaymentTypeEnum[keyof typeof PaymentTypeEnum];
}